refactor: fold storage builders into service
