+++ tests/storage/test_lancedb_store_core.py
@@ -2,7 +2,7 @@ import daft
 import pyarrow as pa
 import pytest
 
-from archetype.app.storage_factory import StorageFactory
+from archetype.app.storage_service import StorageService
 from archetype.core.archetype import Archetype
 from archetype.core.component import Component
 from archetype.core.config import StorageConfig
@@ -76,7 +76,7 @@ async def test_lancedb_store_creates_opens_and_appends(monkeypatch, tmp_path):
 
     monkeypatch.setattr("archetype.core.storage.lancedb.lancedb.connect_async", fake_connect_async)
 
-    uri, namespace = StorageFactory.resolve_location(
+    uri, namespace = StorageService.resolve_location(
         StorageConfig(uri=str(tmp_path / "warehouse"), namespace="ns")
     )
     store = AsyncLancedbStore(uri, namespace)
