+++ tests/storage/test_lancedb_store_errors.py
@@ -2,7 +2,7 @@ import daft
 import pyarrow as pa
 import pytest
 
-from archetype.app.storage_factory import StorageFactory
+from archetype.app.storage_service import StorageService
 from archetype.core.archetype import Archetype
 from archetype.core.component import Component
 from archetype.core.config import StorageConfig
@@ -67,7 +67,7 @@ async def test_lancedb_store_open_table_failure_raises(monkeypatch, tmp_path):
         return client
 
     monkeypatch.setattr("archetype.core.storage.lancedb.lancedb.connect_async", fake_connect_async)
-    uri, namespace = StorageFactory.resolve_location(
+    uri, namespace = StorageService.resolve_location(
         StorageConfig(uri=str(tmp_path / "wh"), namespace="ns")
     )
     store = AsyncLancedbStore(uri, namespace)
@@ -92,7 +92,7 @@ async def test_lancedb_store_append_failure_raises(monkeypatch, tmp_path):
         return client
 
     monkeypatch.setattr("archetype.core.storage.lancedb.lancedb.connect_async", fake_connect_async)
-    uri, namespace = StorageFactory.resolve_location(
+    uri, namespace = StorageService.resolve_location(
         StorageConfig(uri=str(tmp_path / "wh2"), namespace="ns")
     )
     store = AsyncLancedbStore(uri, namespace)
