+++ tests/storage/test_lancedb_store_open_and_optimize.py
@@ -2,7 +2,7 @@ import daft
 import pyarrow as pa
 import pytest
 
-from archetype.app.storage_factory import StorageFactory
+from archetype.app.storage_service import StorageService
 from archetype.core.archetype import Archetype
 from archetype.core.component import Component
 from archetype.core.config import StorageConfig
@@ -67,7 +67,7 @@ class OpenPathClient:
 @pytest.mark.asyncio
 async def test_lancedb_open_path_no_create_called(monkeypatch, tmp_path):
     """When table exists, store should open table and avoid create/index calls."""
-    uri, namespace = StorageFactory.resolve_location(
+    uri, namespace = StorageService.resolve_location(
         StorageConfig(uri=str(tmp_path / "wh"), namespace="ns")
     )
     store = AsyncLancedbStore(uri, namespace)
@@ -137,7 +137,7 @@ async def test_lancedb_optimize_multiple_tables(monkeypatch, tmp_path):
 
     monkeypatch.setattr("archetype.core.storage.lancedb.lancedb.connect_async", fake_connect_async)
 
-    uri, namespace = StorageFactory.resolve_location(
+    uri, namespace = StorageService.resolve_location(
         StorageConfig(uri=str(tmp_path / "wh2"), namespace="ns")
     )
     store = AsyncLancedbStore(uri, namespace)
